breaking: make IOCells aware of the name of their signal
Center naming API around the name of the actual signal being carried by
the IOCell, instead of the name of the cell itself. Expose new
`.named(name)` method on IOCell to signal that it is for a signal named
`name`, and expose getters for name stored on self.

Breaking change to the API of generateFromSignal: now takes name of
signal instead of optional cell name refix